backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
pydantic==2.10.0
pydantic-settings==2.6.1
sqlalchemy==2.0.36
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.28.1
structlog==24.4.0
alembic==1.14.0
slowapi==0.1.9

